Separate out GetProgramName.
