+++ src/common/utility.cpp
@@ -5,7 +5,6 @@
 
 #include "lightstep-tracer-common/collector.pb.h"
 
-#include <unistd.h>
 #include <array>
 #include <cctype>
 #include <cmath>
@@ -41,24 +40,6 @@ timeval ToTimeval(std::chrono::microseconds microseconds) {
   result.tv_usec =
       static_cast<suseconds_t>(num_microseconds % microseconds_in_second);
   return result;
-}
-
-//------------------------------------------------------------------------------
-// GetProgramName
-//------------------------------------------------------------------------------
-std::string GetProgramName() {
-  constexpr int path_max = 1024;
-  std::unique_ptr<char[]> exe_path(new char[path_max]);
-  ssize_t size = ::readlink("/proc/self/exe", exe_path.get(), path_max);
-  if (size == -1) {
-    return "c++-program";  // Dunno...
-  }
-  std::string path(exe_path.get(), size);
-  size_t lslash = path.rfind('/');
-  if (lslash != std::string::npos) {
-    return path.substr(lslash + 1);
-  }
-  return path;
 }
 
 //------------------------------------------------------------------------------
