+++ src/tracer/BUILD
@@ -131,6 +131,7 @@ lightstep_cc_library(
         "tracer.cpp",
     ],
     deps = [
+        "//src/common/system:utility_lib",
         "//src/recorder:auto_recorder_lib",
         "//src/recorder:grpc_transporter_interface",
         "//src/recorder:manual_recorder_lib",
